refactor: Methods and helpers moved to Bullet
