+++ src/Stride.CommunityToolkit/Rendering/ProceduralModels/PrimitiveModelType.cs
@@ -8,15 +8,58 @@ namespace Stride.CommunityToolkit.Rendering.ProceduralModels;
 /// </remarks>
 public enum PrimitiveModelType
 {
+    /// <summary>
+    /// Represents a capsule primitive model.
+    /// </summary>
     Capsule,
+
+    /// <summary>
+    /// Represents a cone primitive model.
+    /// </summary>
     Cone,
+
+    /// <summary>
+    /// Represents a cube primitive model.
+    /// </summary>
     Cube,
+
+    /// <summary>
+    /// Represents a cylinder primitive model.
+    /// </summary>
     Cylinder,
+
+    /// <summary>
+    /// Represents an infinite plane primitive model.
+    /// </summary>
     InfinitePlane,
+
+    /// <summary>
+    /// Represents a plane primitive model.
+    /// </summary>
     Plane,
+
+    /// <summary>
+    /// Represents a rectangular prism primitive model.
+    /// </summary>
     RectangularPrism,
+
+    /// <summary>
+    /// Represents a sphere primitive model.
+    /// </summary>
     Sphere,
+
+    /// <summary>
+    /// Represents a teapot primitive model.
+    /// </summary>
     Teapot,
+
+    /// <summary>
+    /// Represents a torus primitive model.
+    /// </summary>
     Torus,
+
+    /// <summary>
+    /// Represents a triangular prism primitive model.
+    /// </summary>
     TriangularPrism
 }
